[Sampler] Accumulate PyTorch top-p in FP32
`apply_top_k_top_p_pytorch` accumulated the nucleus in the logits dtype.
With a BF16 input the running sum crosses `1 - top_p` too late and extra
low-probability tokens stay in the nucleus: for `linspace(-4, 4)` at
`top_p=0.9` it keeps 2 extra tokens at `V=8192` and 23 at `V=129280`.

`apply_top_k_top_p_triton` asserts float32 input, and the two adjacent
softmax calls in this file already pass `dtype=torch.float32`. This makes
the third one consistent.

Every in-tree caller upcasts to float32 before reaching this helper, so
the defect is latent on main today. It becomes reachable once the
speculative verifier hands head-dtype logits to the sampler.

+++ vllm/v1/sample/ops/topk_topp_sampler.py
@@ -397,7 +397,7 @@ def apply_top_k_top_p_pytorch(
 
     if p is not None:
         # Apply top-p.
-        probs_sort = logits_sort.softmax(dim=-1)
+        probs_sort = logits_sort.softmax(dim=-1, dtype=torch.float32)
         probs_sum = torch.cumsum(probs_sort, dim=-1, out=probs_sort)
         top_p_mask = probs_sum <= 1 - p.unsqueeze(dim=1)
         # at least one
